% Generated by roxygen2 (4.0.2): do not edit by hand
\name{closest.gene.by.interval}
\alias{closest.gene.by.interval}
\title{closest.gene.by.interval}
\usage{
closest.gene.by.interval(noodles)
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look the colses gene}
}
\value{
\code{GRanges} object, noodles argument with added closest gene info for each interval
}
\description{
Find the closest gene for each of th given set of intervals
}

