% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotation.R
\name{closest.gene.start.by.interval}
\alias{closest.gene.start.by.interval}
\title{closest.gene.by.interval}
\usage{
closest.gene.start.by.interval(noodles)
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look the closest gene}
}
\value{
\code{GRanges} object, noodles argument with added closest gene info for each interval
}
\description{
Find the closest gene for each of th given set of intervals
}

