% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CountCoverageOfNoodles}
\alias{CountCoverageOfNoodles}
\title{CountCoverageOfNoodles}
\usage{
CountCoverageOfNoodles(noodles, bedfilnames, bed.ids = bedfilnames)
}
\arguments{
\item{noodles}{\code{GRanges} with the intervals}

\item{bedfilnames}{list of names of bedfiles, one per sample, with some (e.g. methylation) coverage information}

\item{bed.ids}{optional list of names for the samples, they will be used as column names in the result. Default = \code{bedfilnames}}
}
\value{
\code{data.frame}, each row correspond to a noodle; columns are samples
}
\description{
It is a central function of all the Differential.Coverage package. It gets a set of intervals (noodles) as a \code{GRanges} and a list of names of bed files.
Each bedfile represents a sample. For each noodle and each sample, the the sum of the lenght of intervals (read from the sample's bed file) that intersect with the noodle is returned.
}
\seealso{
\code{Differential.Coverage}
}

