% Differential.Coverage is a collection of routines useful for differential analysis of interval (probably, methyl-seq)  
%
% (c) 2014 Alexander Favorov, Elana J. Fertig, Sarah J. Wheelan
%
% favorov@sensi.org
\name{Differential.Coverage}
\alias{Differential.Coverage}
\alias{Differential.Coverage-package}
\docType{package}
\title{
Differential.Coverage package 
}
\description{
\code{Differential.Coverage} is more or less random collection of precedures that are useful for differential analysis of interval sets. Each set represents a genomwide experiment (probably, MBD, but actually, we just suppose it to be an interval genomewide coverage). 
}
