% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gene.list.for.ovelapping.intervals}
\alias{gene.list.for.ovelapping.intervals}
\title{gene.list.for.ovelapping.intervals}
\usage{
gene.list.for.ovelapping.intervals(noodles, flanks = 0)
}
\arguments{
\item{noodles}{the list of intervals to look TSS in}

\item{flanks}{lenght to inflate the noddles by before the search}
}
\value{
\code{GRanges} object, noodles argument with added genes for each interval
}
\description{
Generates list of genes that start inside a given set of intervals
}
\details{
After the noodles (the set of intervals to search TSS in) are inflated by flanks, we look for all the TSS that start inside the (inflated)  intervals according to \code{TxDb.Hsapiens.UCSC.hg19.knownGene}.
If a noodle has any overlapping genes, we form a text list of the genes.
}

