% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{genes.intersected.by.interval}
\alias{genes.intersected.by.interval}
\title{genes.intersected.by.interval}
\usage{
genes.intersected.by.interval(
  noodles,
  flanks = 0,
  genes = NA,
  genome.id = "gencode19"
)
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look TSS in}

\item{flanks}{lenght to inflate the noddles by before the search; if >0, the seqlengths information is to be set in \code{noodles}}

\item{genes}{it is GRandes oject with the annotation genes. The gene_name matadata field is almost required, it passes the gene name to form output. 
The default is NA that means that the function will call \code{\link{get.Known.Gene.List}}}

\item{genome.id}{is genome.annotation.id to call \code{\link{get.Known.Gene.List}}}
}
\value{
\code{GRanges} object, noodles argument with added gene-overlapped genes for each interval
}
\description{
Generates a list of genes (possibly, empty) that intersects with each interval
}
\details{
After the noodles (the set of intervals to search intersections with) are inflated by flanks, we look for all the genes that overlap with the (inflated) intervals. Genes are provided by parameter or they are returned by \code{\link{get.Known.Gene.List}} for \code{genome_id}. If a noodle (interval) overlaps more that one gene, we form a text list of the genes.
}
\seealso{
\code{\link{differential.coverage}}
}
\author{
Alexander Favorov, \email{favorov@sensi.org}
}
\keyword{annotation}
