% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genes.with.TSS.covered.by.interval}
\alias{genes.with.TSS.covered.by.interval}
\title{genes.with.TSS.covered.by.interval}
\usage{
genes.with.TSS.covered.by.interval(noodles, flanks = 0)
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look TSS in}

\item{flanks}{lenght to inflate the noddles by before the search; if >0, the seqlenghtinformation is to be set in \code{noodles}}
}
\value{
\code{GRanges} object, noodles argument with added TSS-overlapped genes for each interval
}
\description{
Generates a list of genes (possibly, empty) that start inside each interval
}
\details{
After the noodles (the set of intervals to search TSS in) are inflated by flanks, we look for all the TSS that start inside the (inflated)  intervals according to \code{TxDb.Hsapiens.UCSC.hg19.knownGene}.
If a noodle (interval) overlaps more that one TSS, we form a text list of the genes.
}

