% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genes.with.TSS.covered}
\alias{genes.with.TSS.covered}
\title{genes.by.TSS.overlap}
\usage{
genes.with.TSS.covered(noodles, flanks = 0)
}
\arguments{
\item{noodles}{the list of intervals to look TSS in}

\item{flanks}{lenght to inflate the noddles by before the search}
}
\value{
\code{GRanges} object that is the list of the genes we look for
}
\description{
Generates list of genes that start inside a given set of intervals
}
\details{
After the noodles (the set of intervals to search TSS in) are inflated by flanks, we look for all the TSS that start inside the (inflated)  intervals according to \code{TxDb.Hsapiens.UCSC.hg19.knownGene}. If the noodles has p.value and/or fdr metadata, we ascribe the data of the interval to the retrieved gene. If there a gene refers to a set of noodles, it has min ascribed. The ishyper data is also transferred to gene, if it is not contradictory.
}

