% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.features.R
\name{get.cytoband.ranges}
\alias{get.cytoband.ranges}
\title{get.cytoband.ranges for hg19, no hg18 or hg 38 version yet}
\usage{
get.cytoband.ranges()
}
\description{
Returns \code{GRanges} with cytobands.
}
\details{
Returns \code{GRanges} with cytobands. Actually, takes \code{\link[SNPchip]{getCytoband}}  from \pkg{SNPchip} and convert it into \code{GRanges}. \code{getCytoband} returns bands in UCSC notation (0-based start, 1-based end), get.cytoband.ranges return it in 1-based.
}
\seealso{
SNPchip, getCytoband
}
