% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get.Known.Gene.List}
\alias{get.Known.Gene.List}
\title{get.Known.Gene.List}
\usage{
get.Known.Gene.List(genome.annotation.id = "gencode19",
  single.strand.genes.only = TRUE)
}
\arguments{
\item{single.strand.genes.only}{seqnames              ranges strand
<Rle>           <IRanges>  <Rle>
96626     chr2 110656009-110664033      +
96626     chr2 111222628-111230652      -}
}
\description{
this thing, we need to unify the TxDb.Hsapiens.UCSC.hg**.knownGene stuff with the gencode data
}
