% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noodles_coverage.R
\name{indicate.any.coverage.of.noodles}
\alias{indicate.any.coverage.of.noodles}
\title{indicate.any.coverage.of.noodles}
\usage{
indicate.any.coverage.of.noodles(noodles, bedfilnames, bed.ids = bedfilnames)
}
\arguments{
\item{noodles}{\code{GRanges} with the intervals}

\item{bedfilnames}{list of names of bedfiles, one per sample, with some (e.g. methylation) coverage information}

\item{bed.ids}{optional list of names for the samples, they will be used as column names in the result. Default = \code{bedfilnames}}
}
\value{
\code{Matrix}, each row correspond to a noodle; columns are samples, sparse=TRUE
}
\description{
It is the alterantive central function of all the differential.coverage package. It gets a set of intervals (noodles) as a \code{GRanges} and a list of names of bed files.
Each bedfile represents a sample. For each noodle and each sample, 0 or 1, which indicates whether the noodle intersects with any bed interval from the sample, is returned.
}
\seealso{
\code{differential.coverage}
}

