% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{inflate.noodles}
\alias{inflate.noodles}
\title{inflate.noodles}
\usage{
inflate.noodles(noodles, flanks = 0, seqlengths = NA)
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to inflate}

\item{flanks}{lenght to inflate the noddles by before the search; if >0, the seqlengths information is to be set as \code{seqlengths} or in \code{seqlengths(noodles)}}

\item{seqlengths}{is to provide the chromosome lentth information without including it in noodles}
}
\description{
Expands each element of GRanges by flanks value both sides. Check chomosome boundaries and avoid breaking them (thus differs from \link{resize}). Needs seqlengths for chromosemes to be defined in GRanges or by seqlengths parameter. If both are given, check whether they do not conradict.
}
