% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chrom.features.R
\name{nucl.chromosomes.hg19}
\alias{nucl.chromosomes.hg19}
\title{nucl.chromosomes.hg19}
\usage{
nucl.chromosomes.hg19(chrM = FALSE)
}
\arguments{
\item{chrM}{boolean: if FALSE (default), returns info only for chr 1..22, X and Y. If TRUE, 1..22, X, Y and M}
}
\value{
\code{seqinfo} object with the chromosome desriptions
}
\description{
Returns the choromosome lengths for hg19
}
\details{
Returns the choromosome lengths for hg19 as a \code{seqinfo} object.
}
\seealso{
seqinfo
}

