% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.noodles.R
\name{prepare.covering.noodles}
\alias{prepare.covering.noodles}
\title{prepare.covering.noodles}
\usage{
prepare.covering.noodles(seq.info, noodle.length = 1000)
}
\arguments{
\item{seq.info}{is \code{seqinfo} or a vector of ints with the names as chromosome.names}

\item{noodle.length}{length of each noodle}
}
\value{
\code{GRanges} that contain the covering set of noodles
}
\description{
prepare intervals (noodles) of the same lenght that adjacently cover a set of chromosomes, which is described by sequinfo
}
\seealso{
\code{GRanges}, \code{seqinfo}
}

