% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulated.fisher.R
\name{prepare.tabulated.fisher}
\alias{prepare.tabulated.fisher}
\title{prepare.tabulated.fisher}
\usage{
prepare.tabulated.fisher(Y, N, folder = ".", load = TRUE, save = TRUE)
}
\arguments{
\item{Y}{cases# , e.g. number of cancer samples}

\item{N}{controls#}

\item{folder}{(default .)}

\item{load}{(default TRUE) logical. If TRUE, we try to load file tabulated.fisher.Y.N.dat}

\item{save}{(default TRUE) logical. If TRUE, we save the \code{tabulated.fisher} to file tabulated.fisher.Y.N.dat}
}
\description{
forms the dataframe (\code{tabulated.fisher}) with the tabulted results for all possible Fisshe exact tests with Y and N numbers of cases and controls
}
\details{
each row gives statistics (p-value, etc) for a 4-pole table:
\tabular{cc}{
MY \tab MN \cr 
Y-MY \tab N-MN }
MY is methylated-positive and MN is methylated-negative (normal)
we want to represent all the Fishers as a (Y+1) rows * (N+1) cols
matrix.
MY=0..Y
MN=0..N
then, we map it to 1..(MY+1)(MN+1), and it is the row index
}

