% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{closest.gene.by.interval}
\alias{closest.gene.by.interval}
\title{closest.gene.by.interval}
\usage{
closest.gene.by.interval(noodles, genome.id = "hg19")
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look TSS in}

\item{genome.id}{is a deprecated alias to genome.annotation.id}
}
\value{
\code{GRanges} object, noodles argument with added closest gene info for each interval
}
\description{
Find the closest gene for each of the given set of intervals
}
