% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{closest.gene.start.by.interval}
\alias{closest.gene.start.by.interval}
\title{closest.gene.start.by.interval}
\usage{
closest.gene.start.by.interval(noodles, genome.id = "hg19")
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look TSS in}

\item{genome.id}{is genome.annotation.id to call \code{\link{get.Known.Gene.List}}}
}
\value{
\code{GRanges} object, noodles argument with added closest gene info for each interval
}
\description{
Find the closest gene start for each of the given set of intervals
}
