% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.coverage.R
\docType{package}
\name{differential.coverage}
\alias{differential.coverage}
\title{differential.coverage: is a collection of routines useful for differential analysis of intervals}
\description{
\code{differential.coverage} is more or less random collection of procedures that are useful for differential analysis of interval sets. 
Each set represents a genomwide experiment (probably, MBD, but actually, we just suppose it to be an interval genomewide coverage). 
The package provides three classes of functions: prepare probes, test differntial intersection of probes with a collection of intervals, and annotation.
}
\section{Preparing probes}{


The probes is a set of genomic intervels. It actually can be some pre-given set of genomic intervals (e.g. open chromatin intervals or CpG islands), but if we want to be agnostic, we can prepare a
uniform tiling of all the genome. We also can 'inflate' all the probes (intervals).\cr
\link{prepare.covering.noodles}\cr
\link{inflate.noodles}\cr
}

\section{Test differential intersection of the prbes with a set of intervals}{


The probes we have is a common gemtry for all the samples we work with. Also, for each sample, we have set of genomic intervals (lets' call the peaks), e.g. MACS results for a MBDseq,
and the samples belong to (two) conditions and for each probe we want to test null hypethesis that the probe interval is equally prone to be intersected with the peaks.\cr
\link{count.coverage.of.noodles}\cr
\link{indicate.any.coverage.of.noodles}\cr
\link{max.peak.score.for.each.noodle}\cr
\link{prepare.tabulated.fisher}\cr
\link{tab.fisher.row.no}\cr
}

