% differential.coverage is a collection of routines useful for differential analysis of interval (probably, methyl-seq)  
%
%
% favorov@sensi.org
\name{differential.coverage}
\alias{differential.coverage}
\alias{differential.coverage-package}
\docType{package}
\title{
differential.coverage package 
}
\description{
\code{differential.coverage} is more or less random collection of precedures that are useful for differential analysis of interval sets. Each set represents a genomwide experiment (probably, MBD, but actually, we just suppose it to be an interval genomewide coverage). 
}
