% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{genes.intersected.by.interval}
\alias{genes.intersected.by.interval}
\title{genes.intersected.by.interval}
\usage{
genes.intersected.by.interval(noodles, flanks = 0, genome.id = "hg19")
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look TSS in}

\item{flanks}{lenght to inflate the noddles by before the search; if >0, the seqlengths information is to be set in \code{noodles}}

\item{genome.id}{is a deprecated alias to genome.annotation.id}
}
\value{
\code{GRanges} object, noodles argument with added gene-overlapped genes for each interval
}
\description{
Generates a list of genes (possibly, empty) that intersects with each interval
}
\details{
After the noodles (the set of intervals to search intersections with) are inflated by flanks, we look for all the genes that overlap with the (inflated)  intervals according to \code{TxDb} object we use (TxDb.Hsapiens.UCSC.hg19.knownGene for genome.id=='hg19', TxDb.Hsapiens.UCSC.hg18.knownGene for hg18 or TxDb.Hsapiens.UCSC.hg38.knownGene for hg38). 
If a noodle (interval) overlaps more that one gene, we form a text list of the genes.
}
