% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{genes.intersected}
\alias{genes.intersected}
\title{genes.intersected}
\usage{
genes.intersected(noodles, flanks = 0, genes = NA, genome.id = "gencode19")
}
\arguments{
\item{noodles}{the \code{GRanges} list of intervals to look TSS in}

\item{flanks}{lenght to inflate the noddles by before the search; if >0, the seqlengths information is to be set in \code{noodles}}

\item{genes}{it is GRandes oject with the annotation genes. The gene_name matadata field is almost required, it passes the gene name to form output. 
The default is NA that means that the function will call \code{\link{get.Known.Gene.List}}}

\item{genome.id}{is genome.annotation.id to call \code{\link{get.Known.Gene.List}}}
}
\value{
\code{GRanges} object that is the list of the genes we look for - the object is not co-indexed with \code{noodles} parameter
}
\description{
Generates list of genes that intersect a given set of intervals
}
\details{
After the noodles (the set of intervals to search intersection with) are inflated by flanks, we look for all the genes that intersect the (inflated) intervals. Genes are provided by parameter or they are returned by \code{\link{get.Known.Gene.List}} for \code{genome_id}. If the noodles has p.value and/or fdr metadata, we ascribe the data of the interval to the retrieved gene. If there a gene refers to a set of noodles, it has min ascribed. The ishyper data is also transferred to gene, if it is not contradictory.
}
\seealso{
\code{\link{differential.coverage}}
}
\author{
Alexander Favorov, \email{favorov@sensi.org}
}
\keyword{annotation}
